# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: La Scarpetta Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List all menu categories
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            default: ru
      responses:
        "200":
          description: List of menu categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"
    post:
      operationId: createMenuCategory
      tags: [admin, menu]
      summary: Create a menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
  /menu/categories/{id}:
    put:
      operationId: updateMenuCategory
      tags: [admin, menu]
      summary: Update a menu category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
    delete:
      operationId: deleteMenuCategory
      tags: [admin, menu]
      summary: Delete a menu category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu/dishes:
    get:
      operationId: listDishes
      tags: [menu]
      summary: List all dishes, optionally filtered by category
      parameters:
        - in: query
          name: categoryId
          schema:
            type: ["integer", "null"]
        - in: query
          name: lang
          schema:
            type: string
            default: ru
        - in: query
          name: featured
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"
    post:
      operationId: createDish
      tags: [admin, menu]
      summary: Create a dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
  /menu/dishes/{id}:
    get:
      operationId: getDish
      tags: [menu]
      summary: Get a dish by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
    put:
      operationId: updateDish
      tags: [admin, menu]
      summary: Update a dish
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
    delete:
      operationId: deleteDish
      tags: [admin, menu]
      summary: Delete a dish
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu/featured:
    get:
      operationId: getFeaturedDishes
      tags: [menu]
      summary: Get featured/signature dishes for homepage
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            default: ru
      responses:
        "200":
          description: Featured dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"

  # ──────────────────── GALLERY ────────────────────
...
  /reservations:
    get:
      operationId: listReservations
      tags: [admin, reservations]
      summary: List all table reservations (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [pending, confirmed, cancelled, completed, null]
        - in: query
          name: date
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Submit a table reservation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [admin, reservations]
      summary: Get reservation by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    patch:
      operationId: updateReservationStatus
      tags: [admin, reservations]
      summary: Update reservation status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List approved reviews
      parameters:
        - in: query
          name: limit
          schema:
            type: ["integer", "null"]
        - in: query
          name: approved
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a customer review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}/approve:
    patch:
      operationId: approveReview
      tags: [admin, reviews]
      summary: Approve or reject a review (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApproval"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    MenuCategory:
      type: object
      required: [id, slug, nameRu, nameIt, nameEn, nameFr, nameZh, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        sortOrder:
          type: integer
    MenuCategoryInput:
      type: object
      required: [slug, nameRu, nameIt, nameEn, nameFr, nameZh]
      properties:
        slug:
          type: string
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        sortOrder:
          type: integer
    Dish:
      type: object
      required: [id, categoryId, nameRu, nameIt, nameEn, nameFr, nameZh, descRu, descIt, descEn, descFr, descZh, price, currency, imageUrl, featured, available]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        available:
          type: boolean
        allergens:
          type: ["string", "null"]
    DishInput:
      type: object
      required: [categoryId, nameRu, nameIt, nameEn, nameFr, nameZh, descRu, descIt, descEn, descFr, descZh, price]
      properties:
        categoryId:
          type: integer
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        available:
          type: boolean
        allergens:
          type: ["string", "null"]

    # ── Gallery ──
...
    Reservation:
      type: object
      required: [id, name, phone, date, time, guests, status, lang, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        comment:
          type: ["string", "null"]
        lang:
          type: string
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    ReservationInput:
      type: object
      required: [name, phone, date, time, guests]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        comment:
          type: ["string", "null"]
        lang:
          type: string
    ReservationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        adminNotes:
          type: ["string", "null"]
...
    Review:
      type: object
      required: [id, authorName, rating, text, lang, approved, createdAt]
      properties:
        id:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        text:
          type: string
        lang:
          type: string
        source:
          type: ["string", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [authorName, rating, text]
      properties:
        authorName:
          type: string
        rating:
          type: integer
        text:
          type: string
        lang:
          type: string
        source:
          type: ["string", "null"]
    ReviewApproval:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
...