# lib/api-zod/src/generated/types/dishInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * La Scarpetta Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface DishInput {
  categoryId: number;
  nameRu: string;
  nameIt: string;
  nameEn: string;
  nameFr: string;
  nameZh: string;
  descRu: string;
  descIt: string;
  descEn: string;
  descFr: string;
  descZh: string;
  price: number;
  currency?: string;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  available?: boolean;
  /** @nullable */
  allergens?: string | null;
}