# lib/api-zod/src/generated/types/reservationStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * La Scarpetta Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationStatusUpdateStatus } from './reservationStatusUpdateStatus';

export interface ReservationStatusUpdate {
  status: ReservationStatusUpdateStatus;
  /** @nullable */
  adminNotes?: string | null;
}